refactor : 엔티티, 레포지토리 위치 조정
- Address, Image, Video 엔티티와 레포지토리를 capsule 하위로 이동
