+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/repository/AddressRepository.java
@@ -1,7 +1,7 @@
-package site.timecapsulearchive.core.global.repository;
+package site.timecapsulearchive.core.domain.capsule.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.timecapsulearchive.core.global.entity.Address;
+import site.timecapsulearchive.core.domain.capsule.entity.Address;
 
 public interface AddressRepository extends JpaRepository<Address, Long> {
