+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/history/entity/HistoryImage.java
@@ -13,7 +13,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import site.timecapsulearchive.core.global.entity.BaseEntity;
-import site.timecapsulearchive.core.global.entity.Image;
+import site.timecapsulearchive.core.domain.capsule.entity.Image;
 
 @Entity
 @Getter
